move 'SyndicateAccountInfo' into its unique api route file
